requirements.txt
streamlit
PyPDF2==3.0.1
google.generativeai
python-dotenv
streamlit_extras
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
google-generativeai==0.3.0
python-dotenv==1.0.0
pymongo==4.6.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dateutil==2.8.2
email-validator>=2.1.1